feat: metadata 관련 float -> double로 변경

+++ src/main/java/com/threestar/trainus/domain/lesson/student/dto/LessonSearchResponseDto.java
@@ -11,7 +11,7 @@ public record LessonSearchResponseDto(
 	String lessonLeaderName,
 	String lessonLeaderImage,
 	int reviewCount,
-	float rating,
+	double rating,
 	Category category,
 	int price,
 	int maxParticipants,
